Extend sensor name support

+++ core/src/main/java/cn/edu/tsinghua/iot/benchmark/workload/RealDataWorkLoad.java
@@ -36,7 +36,7 @@ public class RealDataWorkLoad extends DataWorkLoad {
 
   public RealDataWorkLoad(List<String> files) {
     dataReader = DataReader.getInstance(files);
-    batchNumber = files.size() * config.getBIG_BATCH_SIZE();
+    batchNumber = (long) files.size() * config.getBIG_BATCH_SIZE();
   }
 
   @Override
